test(AvlTree): split tests for AvlTree for clarity
> Move tests for joining AvlTree-s to separate file
> Add helper function reused by both main tests for AvlTree and joinAvlTrees
